Change output screen size callback to configuration parameter
This changes the behaviour as to absolute events and touch events to
only be in the output screen coordinate space if an output screen size
has been set.

If an output screen size has not been set, the coordinates of the
mentioned events will be in device specific coordinate space.

Signed-off-by: Jonas Ådahl <jadahl@gmail.com>