evdev: Define KEY_LIGHTS_TOGGLE macro if missing
When building on a system with an older kernel, some KEY_ macros might
be missing. To be able to build on such system, define them if they are
missing.

It is probably better to keep our own copy of input.h somewhere in our
tree, and include that one instead of the system one, but that can be
added later.

Signed-off-by: Jonas Ådahl <jadahl@opera.com>